requirements.txt
amqp==5.1.1
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.2
backcall==0.2.0
bcolz==1.2.1
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.7
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==41.0.1
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Deprecated==1.2.14
easydict==1.10
entrypoints==0.4
exceptiongroup==1.1.2
fastapi==0.81.0
fastapi-event==0.1.3
filelock==3.12.2
fonttools==4.38.0
fsspec==2023.1.0
gitdb==4.0.10
GitPython==3.1.31
graphviz==0.20.1
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
httptools==0.6.0
huggingface-hub==0.16.4
idna==3.4
importlib-metadata==6.7.0
ipykernel==6.16.2
ipython==7.34.0
jedi==0.18.2
joblib==1.3.1
jupyter_client==7.4.9
jupyter_core==4.12.0
jwt==1.3.1
kiwisolver==1.4.4
kombu==5.2.4
matplotlib==3.5.3
matplotlib-inline==0.1.6
mxnet==1.2.0
nest-asyncio==1.5.6
numpy==1.21.6
opencv-python==4.8.0.74
packaging==23.1
pandas==1.3.5
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
prompt-toolkit==3.0.39
protobuf==4.23.4
psutil==5.9.5
pycparser==2.21
pydantic==1.10.2
pydantic_core==2.1.2
Pygments==2.15.1
PyJWT==2.7.0
pymssql==2.2.7
pyodbc==4.0.39
pyparsing==3.1.0
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.6
pythondi==1.2.4
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.1.0
redis==4.3.4
requests==2.31.0
safetensors==0.3.1
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
SQLAlchemy==2.0.18
starlette==0.19.1
tensorboardX==2.6.1
threadpoolctl==3.1.0
timm==0.9.2
torch==1.13.1+cu116
torchaudio==0.13.1+cu116
torchvision==0.14.1+cu116
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.7.1
ujson==5.4.0
ultralytics==8.0.131
urllib3==2.0.3
uvicorn==0.22.0
vine==5.0.0
watchfiles==0.19.0
wcwidth==0.2.6
websockets==11.0.3
wrapt==1.15.0
zipp==3.15.0
